TEK-528
Trying to avoid the bug that shows up if a product has a compare at price that is the same as the price on the same variant and no other variant with a price lower than the compare at price. Bug results in an infinite number for the discount and the sale price to 1 million nok.